move CachedFileDownloader to downloaders module
